requirements_ubuntu.txt
absl-py==2.4.0
accelerate==1.10.1
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.2
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
autopage==0.6.0
av==15.1.0
azure-core==1.34.0
azure-devops==7.1.0b4
backports.strenum==1.3.1
builtin-interfaces==1.2.1
cachetools==6.2.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cliff==4.13.2
cmaes==0.12.0
coloredlogs==15.0.1
colorlog==6.10.1
composition-interfaces==1.2.1
contourpy==1.3.2
coverage==7.13.4
cryptography==46.0.5
cv-bridge==3.2.1
cycler==0.12.1
Cython==3.0.12
datasets==3.5.1
decorator==4.4.2
demo-nodes-py==0.20.5
diagnostic-msgs==4.8.0
dill==0.3.8
dlib==20.0.0
docker==7.1.0
docker-pycreds==0.4.0
domain-coordinator==0.10.0
editdistance==0.8.1
einops==0.8.2
evdev==1.9.2
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.3.1
face-recognition==1.3.0
face_recognition_models==0.3.0
ffmpeg-python==0.2.0
filelock==3.25.1
flatbuffers==25.12.19
fonttools==4.62.0
frozenlist==1.8.0
fsspec==2025.3.0
future==1.0.0
gast==0.7.0
geometry-msgs==4.8.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.49.0
google-auth-oauthlib==1.3.0
google-pasta==0.2.0
googleapis-common-protos==1.73.0
greenlet==3.2.3
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.11
iglovikov-helper-functions==0.0.53
image-geometry==3.2.1
imagecorruptions==1.1.2
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
iniconfig==2.1.0
interactive-markers==2.3.2
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jpeg4py==0.1.4
julius==0.2.7
keras==3.12.1
kiwisolver==1.5.0
laser-geometry==2.4.0
launch==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
lazy-loader==0.5
libclang==18.1.1
librosa==0.11.0
lifecycle-msgs==1.2.1
lightning==2.5.3
lightning-utilities==0.15.3
llvmlite==0.46.0
logging-demo==0.20.5
Mako==1.3.10
map-msgs==2.1.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.32
message-filters==4.3.6
ml_dtypes==0.5.4
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
msrest==0.7.1
multidict==6.7.1
multiprocess==0.70.16
namex==0.1.0
natsort==8.4.0
nav-msgs==4.8.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.64.0
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.16.1
onnxruntime-gpu==1.19.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.9.0.80
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optree==0.19.0
optuna==4.7.0
optuna-integration==4.5.0
osrf-pycommon==2.1.6
packaging==26.0
pandas==2.2.3
pbr==6.1.1
pcl-msgs==1.0.0
peft==0.10.0
pendulum-msgs==0.20.5
pillow==12.1.1
pip==26.0.1
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.9.0
praat-parselmouth==0.4.5
prettytable==3.17.0
primePy==1.3
proglog==0.1.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.0.0
pyannote-audio==4.0.4
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.4.0
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
Pympler==1.1
pynput==1.7.7
pyparsing==3.3.2
pyperclip==1.11.0
pysndfx==0.3.6
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-qt-binding==1.1.2
python_speech_features==0.6
python-xlib==0.33
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.9.0
pytz==2026.1.post1
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
resource-retriever==3.1.3
retinaface-pytorch==0.0.1
rich==14.0.0
rich-argparse==1.7.2
rmw-dds-common==1.6.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.4
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.2
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
sentencepiece==0.2.1
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==80.10.2
sgmllib3k==1.0.0
shape-msgs==4.8.0
shellingham==1.5.4
simsimd==6.5.13
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
sphfile==1.0.3
SQLAlchemy==2.0.41
sros2==0.10.6
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
stevedore==5.7.0
stringzilla==4.6.0
sympy==1.14.0
tabulate==0.9.0
teleop-twist-keyboard==2.4.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.21.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
tf-keras==2.15.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
timm==0.9.7
tokenizers==0.13.3
tomli==2.2.1
topic-monitor==0.20.5
torch-adopt==0.1.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchcodec==0.10.0
torchinfo==1.8.0
torchmetrics==1.9.0
tqdm==4.67.3
trajectory-msgs==4.8.0
transformers==4.30.0
transforms==0.2.1
triton==3.6.0
turtlesim==1.4.2
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.0
tzdata==2025.2
unique-identifier-msgs==2.2.1
urllib3==2.6.3
visualization-msgs==4.8.0
wandb==0.19.10
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
xxhash==3.6.0
yacs==0.1.8
yarl==1.23.0
zipp==3.23.0
